1. made the two controllers each control their own collision in their world 2. made collisioncontroller work on pick up 3. everything is normal except i took out trash
